+++ src/main/java/com/journaldev/spring/dao/CartDAOImpl.java
@@ -1,5 +1,6 @@
 package com.journaldev.spring.dao;
  
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.hibernate.Query;
@@ -119,13 +120,13 @@ public class CartDAOImpl implements CartDAO {
     }
     
     /*@Override*/
-    public int getCartItemCostByUserLogin(String name) {
+    public BigDecimal getCartItemCostByUserLogin(String name) {
         Session session = this.sessionFactory.getCurrentSession();      
         String hql = "SELECT quantity * SUM(pricePerItem) FROM Cart WHERE accountId = ("
         		+ "SELECT accountId FROM Account where userLogin = :userLogin)";
         Query query = session.createQuery(hql);
         query.setParameter("userLogin", name);
-        int cartSum = (int) query.uniqueResult();
+        BigDecimal cartSum = (BigDecimal) query.uniqueResult();
 
         logger.info("getCartByUserLogin query: " + query.toString());
         logger.info("getCartByUserLogin query results (toString()): " + cartSum);        
@@ -133,13 +134,13 @@ public class CartDAOImpl implements CartDAO {
     }
     
     /*@Override*/
-    public int getCartShippingCostByUserLogin(String name) {
+    public BigDecimal getCartShippingCostByUserLogin(String name) {
         Session session = this.sessionFactory.getCurrentSession();      
         String hql = "SELECT quantity * SUM(shippingCost) FROM Cart WHERE accountId = ("
         		+ "SELECT accountId FROM Account where userLogin = :userLogin)";
         Query query = session.createQuery(hql);
         query.setParameter("userLogin", name);
-        int cartSum = (int) query.uniqueResult();
+        BigDecimal cartSum = (BigDecimal) query.uniqueResult();
 
         logger.info("getCartByUserLogin query: " + query.toString());
         logger.info("getCartByUserLogin query results (toString()): " + cartSum);        
@@ -147,13 +148,13 @@ public class CartDAOImpl implements CartDAO {
     }         
     
     /*@Override*/
-    public int getCartTaxCostByUserLogin(String name) {
+    public BigDecimal getCartTaxCostByUserLogin(String name) {
         Session session = this.sessionFactory.getCurrentSession();      
         String hql = "SELECT quantity * SUM(tax) FROM Cart WHERE accountId = ("
         		+ "SELECT accountId FROM Account where userLogin = :userLogin)";
         Query query = session.createQuery(hql);
         query.setParameter("userLogin", name);
-        int cartSum = (int) query.uniqueResult();
+        BigDecimal cartSum = (BigDecimal) query.uniqueResult();
 
         logger.info("getCartByUserLogin query: " + query.toString());
         logger.info("getCartByUserLogin query results (toString()): " + cartSum);        
@@ -161,13 +162,13 @@ public class CartDAOImpl implements CartDAO {
     }      
     
     /*@Override*/
-    public int getCartTotalByUserLogin(String name) {
+    public BigDecimal getCartTotalByUserLogin(String name) {
         Session session = this.sessionFactory.getCurrentSession();      
         String hql = "SELECT quantity * (SUM(pricePerItem) + SUM(shippingCost) + SUM(tax)) FROM Cart WHERE accountId = ("
         		+ "SELECT accountId FROM Account where userLogin = :userLogin)";
         Query query = session.createQuery(hql);
         query.setParameter("userLogin", name);
-        int cartSum = (int) query.uniqueResult();
+        BigDecimal cartSum = (BigDecimal) query.uniqueResult();
 
         logger.info("getCartByUserLogin query: " + query.toString());
         logger.info("getCartByUserLogin query results (toString()): " + cartSum);
