+++ src/main/java/com/journaldev/spring/service/CartService.java
@@ -1,7 +1,9 @@
 package com.journaldev.spring.service;
  
+import java.math.BigDecimal;
 import java.util.List;
  
+
 
 
 
@@ -19,9 +21,9 @@ public interface CartService {
     public void addToCart(int accountId,int inventoryId,int quantity, String pricePerItem,String shippingCost,String tax);
     public void removeCart(int cartId);
     public List<Cart> getCartByUserLogin(String name);
-    public int getCartItemCostByUserLogin(String name);
-    public int getCartShippingCostByUserLogin(String name);
-    public int getCartTaxCostByUserLogin(String name);    
-    public int getCartTotalByUserLogin(String name);
+    public BigDecimal getCartItemCostByUserLogin(String name);
+    public BigDecimal getCartShippingCostByUserLogin(String name);
+    public BigDecimal getCartTaxCostByUserLogin(String name);    
+    public BigDecimal getCartTotalByUserLogin(String name);
      
 }
